--- launcher/minecraft/auth/AuthStep.cpp
@@ -1,3 +1,0 @@
-#include "AuthStep.h"
-
-AuthStep::AuthStep(AccountData* data) : QObject(nullptr), m_data(data) {}
