+++ launcher/minecraft/auth/flows/AuthFlow.cpp
@@ -4,11 +4,13 @@
 #include <QNetworkRequest>
 
 #include "AuthFlow.h"
-#include "katabasis/Globals.h"
 
 #include <Application.h>
 
-AuthFlow::AuthFlow(AccountData* data, QObject* parent) : AccountTask(data, parent) {}
+AuthFlow::AuthFlow(AccountData* data, QObject* parent) : Task(parent), m_data(data)
+{
+    changeState(AccountTaskState::STATE_CREATED);
+}
 
 void AuthFlow::succeed()
 {
@@ -46,16 +48,26 @@ void AuthFlow::nextStep()
 QString AuthFlow::getStateMessage() const
 {
     switch (m_taskState) {
-        case AccountTaskState::STATE_WORKING: {
-            if (m_currentStep) {
+        case AccountTaskState::STATE_CREATED:
+            return "Waiting...";
+        case AccountTaskState::STATE_WORKING:
+            if (m_currentStep)
                 return m_currentStep->describe();
-            } else {
-                return tr("Working...");
-            }
-        }
-        default: {
-            return AccountTask::getStateMessage();
-        }
+            return tr("Working...");
+        case AccountTaskState::STATE_SUCCEEDED:
+            return tr("Authentication task succeeded.");
+        case AccountTaskState::STATE_OFFLINE:
+            return tr("Failed to contact the authentication server.");
+        case AccountTaskState::STATE_DISABLED:
+            return tr("Client ID has changed. New session needs to be created.");
+        case AccountTaskState::STATE_FAILED_SOFT:
+            return tr("Encountered an error during authentication.");
+        case AccountTaskState::STATE_FAILED_HARD:
+            return tr("Failed to authenticate. The session has expired.");
+        case AccountTaskState::STATE_FAILED_GONE:
+            return tr("Failed to authenticate. The account no longer exists.");
+        default:
+            return tr("...");
     }
 }
 
@@ -63,5 +75,64 @@ void AuthFlow::stepFinished(AccountTaskState resultingState, QString message)
 {
     if (changeState(resultingState, message)) {
         nextStep();
+    }
+}
+
+bool AuthFlow::changeState(AccountTaskState newState, QString reason)
+{
+    m_taskState = newState;
+    // FIXME: virtual method invoked in constructor.
+    // We want that behavior, but maybe make it less weird?
+    setStatus(getStateMessage());
+    switch (newState) {
+        case AccountTaskState::STATE_CREATED: {
+            m_data->errorString.clear();
+            return true;
+        }
+        case AccountTaskState::STATE_WORKING: {
+            m_data->accountState = AccountState::Working;
+            return true;
+        }
+        case AccountTaskState::STATE_SUCCEEDED: {
+            m_data->accountState = AccountState::Online;
+            emitSucceeded();
+            return false;
+        }
+        case AccountTaskState::STATE_OFFLINE: {
+            m_data->errorString = reason;
+            m_data->accountState = AccountState::Offline;
+            emitFailed(reason);
+            return false;
+        }
+        case AccountTaskState::STATE_DISABLED: {
+            m_data->errorString = reason;
+            m_data->accountState = AccountState::Disabled;
+            emitFailed(reason);
+            return false;
+        }
+        case AccountTaskState::STATE_FAILED_SOFT: {
+            m_data->errorString = reason;
+            m_data->accountState = AccountState::Errored;
+            emitFailed(reason);
+            return false;
+        }
+        case AccountTaskState::STATE_FAILED_HARD: {
+            m_data->errorString = reason;
+            m_data->accountState = AccountState::Expired;
+            emitFailed(reason);
+            return false;
+        }
+        case AccountTaskState::STATE_FAILED_GONE: {
+            m_data->errorString = reason;
+            m_data->accountState = AccountState::Gone;
+            emitFailed(reason);
+            return false;
+        }
+        default: {
+            QString error = tr("Unknown account task state: %1").arg(int(newState));
+            m_data->accountState = AccountState::Errored;
+            emitFailed(error);
+            return false;
+        }
     }
 }
